TEA: attempt to detect open/closed hand

+++ BossMod/Modules/Shadowbringers/Ultimate/TEA/P1HandOfPain.cs
@@ -12,7 +12,7 @@ namespace BossMod.Shadowbringers.Ultimate.TEA
             var hand = module.Enemies(OID.LiquidHand).FirstOrDefault();
             if (boss != null && hand != null)
             {
-                var diff = (hand.HP.Cur - boss.HP.Cur) * 100.0f / boss.HP.Max;
+                var diff = (int)(hand.HP.Cur - boss.HP.Cur) * 100.0f / boss.HP.Max;
                 hints.Add($"Hand HP: {(diff > 0 ? "+" : "")}{diff:f1}%");
             }
         }
